Add workaround to fix value of cumulative_clinical_episodes_by_location_age_ in MDC: 1. Change long vector to std::vector<unsigned int> 2. remove last_10_fraction_positive_that_are_clinical_by_location_age_class_by_5_ in MDC

+++ src/Utils/TypeDef.h
@@ -20,7 +20,7 @@ class IStrategy;
 
 class Therapy;
 
-using Ul = uint64_t;
+using Ul = unsigned int;
 
 using DoubleVector = std::vector<double>;
 using DoubleVector2 = std::vector<DoubleVector>;
